refactor(i18n): reorganize translation files and update paths
Move machine-translated files from translate/ to locales/ directory
Update README location and clarify translation maintenance status
