+++ src/renderer/src/i18n/index.ts
@@ -4,18 +4,16 @@ import { defaultLanguage } from '@shared/config/constant'
 import i18n from 'i18next'
 import { initReactI18next } from 'react-i18next'
 
-// Original translation
+import deDE from './locales/de-de.json'
+import elGR from './locales/el-gr.json'
 import enUS from './locales/en-us.json'
+import esES from './locales/es-es.json'
+import frFR from './locales/fr-fr.json'
+import jaJP from './locales/ja-jp.json'
+import ptPT from './locales/pt-pt.json'
+import ruRU from './locales/ru-ru.json'
 import zhCN from './locales/zh-cn.json'
 import zhTW from './locales/zh-tw.json'
-// Machine translation
-import deDE from './translate/de-de.json'
-import elGR from './translate/el-gr.json'
-import esES from './translate/es-es.json'
-import frFR from './translate/fr-fr.json'
-import jaJP from './translate/ja-jp.json'
-import ptPT from './translate/pt-pt.json'
-import ruRU from './translate/ru-ru.json'
 
 const logger = loggerService.withContext('I18N')
