+++ src/renderer/src/i18n/locales/README.md
