+++ src/renderer/src/i18n/locales/de-de.json
