+++ src/renderer/src/i18n/locales/el-gr.json
