+++ src/renderer/src/i18n/locales/es-es.json
