+++ src/renderer/src/i18n/locales/fr-fr.json
