+++ src/renderer/src/i18n/locales/ja-jp.json
