+++ src/renderer/src/i18n/locales/pt-pt.json
