+++ src/renderer/src/i18n/locales/ru-ru.json
